backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
supabase==2.0.3
pinecone-client==3.0.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.6.1
pymupdf==1.23.8
sentence-transformers>=2.2.2
torch>=2.1.0
tiktoken==0.5.2
httpx>=0.24.0,<0.25.0
pydantic==2.5.2
python-multipart==0.0.6
